requirements.txt
pandas
numpy
schedule
litellm
openai
langchain
langchain-openai
python-dotenv
tabulate
colorama